[#18] refactor: sns login 버튼 컴포넌트 디렉토리 생성
